# arxflow.client/src/api/generated/model/createContraparteRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * ArxFlow API
 * OpenAPI spec version: v1
 */

export interface CreateContraparteRequest {
  /** @nullable */
  nome: string | null;
}