# arxflow.client/src/api/generated/model/createFundoRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * ArxFlow API
 * OpenAPI spec version: v1
 */

export interface CreateFundoRequest {
  /** @nullable */
  nome: string | null;
  /** @nullable */
  cnpj?: string | null;
  /** @nullable */
  alphaToolsId?: string | null;
}