# arxflow.client/src/api/generated/model/updateEmissorRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * ArxFlow API
 * OpenAPI spec version: v1
 */

export interface UpdateEmissorRequest {
  /** @nullable */
  nome: string | null;
  /** @nullable */
  documento?: string | null;
  /** @nullable */
  alphaToolsId?: string | null;
}